+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Format `<github issue/pr number>: <short description>`.
 
 ## SNAPSHOT
 
+* [#1927](https://github.com/kroxylicious/kroxylicious/issues/1927): chore(docs): Remove deprecated adminHttp configuration property.
 * [#1885](https://github.com/kroxylicious/kroxylicious/issues/1885): chore(docs): Remove deprecated support for virtualClusters expressed as a map.
 * [#2598](https://github.com/kroxylicious/kroxylicious/pull/2598): feat(metrics): Add metrics for the number of active connections
 
@@ -14,6 +15,8 @@ Format `<github issue/pr number>: <short description>`.
 
 * The `virtualClusters` configuration property now requires a list of `virtualCluster` objects.  The support
   for supplying a virtual cluster map (which was deprecated at 0.11.0) is now removed.
+* Support for the `adminHttp` configuration property (which was deprecated in 0.11.0) is removed. Use `management` instead.
+  Also support for the `host` configuration property within that object (which was also deprecated in 0.11.0) is removed. Use `bindAddress` instead.
 
 ## 0.15.0